Implements `args` command
When `args` command is run, output will be the arguments necessary for bindings to be passed in correctly to either `pack build`, if you set the `-p` flag, or `docker run`, if you set the `-d` flag. This ends up setting a volume mount to the local bindings directory and mapping that to `/bindings`. It also sets the `SERVICE_BINDING_ROOT` env variable so that the bindings are found at build/run time.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>