Implement 'ca-certs' subcommand
This takes one or more '--cert' arguments, which point to a certificate file. The certificate files are added as a binding of type `ca-certificates` and the binding files assume the file name of the certificate (i.e. what's after the last /), or if it cannot determine the file name then it'll use 'cert-X' where X is the argument number.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>