Do not quote volumes
This is causing problems when running as a subshell to pass arguments through to pack. Not sure why, as it works when I manually use them. For now, this is working around the problem. It has the drawback that it'll fail if there are unquoted values in the directory (like spaces).

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>