Switch subcommand args to init
Outtputing the list of args to include is a bit problematic. It works OK with bash, but there's a bug in fish shell that makes this really awkward. It also still requires you to have some sort of alias or wrapper function to make the integration smooth.

Instead of that, we're switching to an init command which just outputs a couple of wrapper functions, one for fish shell and one for bash shell. The wrapper scripts will overlay the docker and pack commands. If a `pack build` or `docker run` is executed, the script will run that command but with the additional arguments appended. If any other `docker` or `pack` command is run, the arguments are passed through unmodified.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>