Fix timeouts & document settings
Fixes the usage of timeouts with the download client. Previously, a timeout on the entire request was being set to 30s. In some cases, you could go over this timeout and it would cause a download failure. Now the default timeouts are a connection timeout of 5secs and a read timeout of 5secs. This means that you have up to 5 seconds to make a TCP connection and you have up to 5 seconds for each individual read on the socket.

For backward compatibility, you may still configure the `BT_REQ_TIMEOUT` value which will set the overall request timeout. This is not recommended though. It defaults to being unset, which does not enforce an overall timeout on the request.

In addition, this commit includes documentation for each of the environment variables you can set to configure the client.

Resolves #3.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>